+++ serverless/src/utils.ts
@@ -15,7 +15,7 @@ export function createPresignedURL(
     secretAccessKey = process.env.AWS_SECRET_ACCESS_KEY,
     sessionToken = process.env.AWS_SESSION_TOKEN,
     // expires = 0, // @TODO: 300, check if this is working http://docs.aws.amazon.com/AmazonS3/latest/API/sigv4-query-string-auth.html
-  } = {}
+  } = {},
 ): string {
   const signed = aws4.sign(
     {
@@ -31,11 +31,11 @@ export function createPresignedURL(
     {
       accessKeyId,
       secretAccessKey,
-    }
+    },
   )
 
   return `wss://${host}${signed.path}&X-Amz-Security-Token=${encodeURIComponent(
-    sessionToken
+    sessionToken,
   )}`
 }